requirements.txt
streamlit
opencv-contrib-python
numpy
pandas
pillow